rpmcpio.h: change fnamelen to size_t
Otherwise there is a hole in the structure, which makes me
uncomfortable. :-)

+++ rpmcpio.h
@@ -63,7 +63,7 @@ struct cpioent {
 	unsigned long long linklen;
     };
     // Filename length: fnamelen = strlen(fname) < PATH_MAX = 4096.
-    unsigned fnamelen;
+    size_t fnamelen;
     // The filename of the entry, null-terminated.  Source packages have
     // basename-only filenames with no slashes in them.  Binary packages have
     // absolute pathnames which start with '/'.
